// frontend/src/api/bindings/AgendaElectKind.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Auto-generated discriminant enum variants
 */
export type AgendaElectKind = "ForOrAgainst" | "Player" | "StrategyCard" | "Law" | "SecretObjective" | "Planet" | "PlanetWithTrait" | "CulturalPlanet" | "HazardousPlanet" | "IndustrialPlanet";